Recreated Database & updated FrontEnd
Sometimes coding is too annoing, thats why i remembered about SOLID principles and cut a little code where it was needed.

In this version i additionally recreated a database due to previous one was too awfut to work with. It just didn't corresponded to needed features implementation.

What about this? I do not know, i will try to find if the new is working better, but for now it seems just fine.