Lecturer Page the very beginning
In this commit fixed bag with showing lecturer and student pages tables.

Additionally there was implemented methods for lecturer that already workd, but no correctly in the end due to the customer/user design logic issues i faced.

In next commit will be updated Admin buttons and methods. Will be removed button that adds lecturer to course and will be created button that assigns subjects to lecturer.
This is loggicaly more proper way to store information about lecturer enrolement in course and wgat subjects does lecturer has that leads to what subjects course has!