Final version?
Yep, i did all i could. I just can not continue development of an app due to i ran out of time!

There is still a lot could be done. Can be added: validations features, some user interactions features and so on...

I hope this is the last commit, but everithing can change and in the bad case i will force my self additionally to rewrite code with useing SOLID principles, if they are missed there. I don't remember have i implemented all of them, i kust wrote alot of code, so maybe there will be some due to it is a factor of habbit to use SOLID during development not even yhinking about it